+++ code/go/0chain.net/smartcontract/storagesc/readpool.go
@@ -138,7 +138,7 @@ func (rp *readPool) moveToBlobber(sscKey, allocID, blobID string,
 			"allocation: %s, blobber: %s", allocID, blobID)
 	}
 
-	err = sp.DistributeRewards(float64(value), blobID, spenum.Blobber, balances)
+	err = sp.DistributeRewards(value, blobID, spenum.Blobber, balances)
 	if err != nil {
 		return "", fmt.Errorf("can't move tokens to blobber: %v", err)
 	}
